UI updates and major features
- updated design and fixed draggable element issues.
- simplified and improved styling
- extra functionality for starting and pausing camera
- App works offline

Specifically for this commit, improvements have been made to the process of starting and managing the camera, which has also involved some UI changes. 
Processing the frames has started to be moved into a WebWorker (separate thread) which has invloved the use of an OffscreenCanvas for rendering the ImageBitmap data.