Character Segmentation
Individual drawings now displayed in separate container.

All the thresholding is now done in the getFrameThread, reducing load on main thread. 
getFrameThread now sends a characterDrawn event to main thread when there is no movement and the user has drawn 'enough'. Main thread clones svg container and puts clone in the main container. 
Uses the min and max points from the polyline as the start x and y coordinates to position the polyline in the clone. 

Toolbar area has been reduced and now covers the drawing container. This was because the toolbar container filled the whole viewport blocking touch inputs. It was moved onto the drawing container as this container didn't have any touch inputs. Now, drawn characters are placed in the main container. This container is the only container that can scroll, allowing extra drawn characters to be revealed.