Server Response
Coordinates of drawn character are sent to server, and overlays the result on the drawn character. 

The scaledX and scaledY coordinates in the getFrameThread are wrapped up in an object which is sent to the server in JSON format. There, simple transformations are applied to the data, similar to those implicitly applied to the cloned SVG drawing when the viewBox is changed (i.e, the datapoints are translated to the min x and re-scaled). Then, the angles between the datapoints are calculated, and compared to two pre-defined series of datapoints for the letters O and X by way of MSE. The character with the lowest MSE is returned to the client, where the character is overlayed on the respective drawing.